Engines: Fix dealing with auxiliary variables in transition systems
After we allowed auxiliary variables in init and query, we encountered
auxiliary variables present in the definitions of predicates in the
model. Such variables would have to be existentially quantified for the
model to be correct, but we don't support that.
In any case, the question whether or not an algorith can handle
auxiliary variables in init adn query is more complicated than it
originally seemed.
For that reason we now mostly revert the implemented change.
We still allow auxiliary variables in the *definition* of transition
system, but the engines still eliminate them from init and query.
For PDKIND, I believe auxiliary variables in init are OK, so we only
eliminate them from the query.
For other engines, the question needs to be revisited later.