Basic implementation of reachable lasso detector
Detecting a reachable lasso means the transition system is
nonterminating, meaning there is an infinite trace.
If there is no lasso, the system is terminating only of the domain is
finite. An infinite state system may allow for an inifinite trace on
which no state is repeated.
We need additional mechanism to detect infinite traces in such case.