Termination: ITS frontend and encoding to Transition System
This adds a prototype for parsing Integer Transition Systems in .ari
files and a simple translation to CHCs.
This gives us a linear CHC system which we then preprocess and
eventually encode as a transition system with single transition
relation.
We can run our lasso detector on such a system and determine
nontermination if a lasso is found.