K- to 1-inductive invariant with sequence interpolants
Based on the advice of Prof. McMillan, the computation of 1-inductive
invariant can be done with a single query using sequence interpolants.
Compared to the previous approach, we need fewer queries (1 vs log k),
but we compute more interpolants (k vs log k).
In general the new approach is significantly faster, but there are some
cases where the validation of the resulting 1-inductive invariant is
slower.

It would be interesting to compare translation + validation time of both
approaches.