Add new backend engine: Transitive Relations Learning
TRL is a new approach for proving safety of infinite-state systems.
It works like BMC, but adds new relations as a possible step, instead of
the original transition relation.
With the new relations, we can effectively shorten the diameter of the
system, possibly making it finite instead of infinite, allowing BMC-like
style exploration to prove absence of bugs.

The implementation is based on the paper by Florian Frohn and Jurgen
Giesl: Infinite State Model Checking by Learning Transitive Relations.

CADE'25: https://dl.acm.org/doi/10.1007/978-3-031-99984-0_25
Extended version: https://arxiv.org/pdf/2502.04761

This is the initial implementation in Golem, which is still missing a
few things.
We need better detection of recurrent literals, including divisibility
constraints, a way to check error traces and possibly a way to refine
relations based on a spurious counterexample.