TPA: Fix engine reset
TPA engine has a method to reset it to solve a new transition system.
However, this has never been used before in an incremental way.
It actually did not clear the engine completely, it kept around the
reachability solvers. Thus, trying to reset it with a new transition
system did not work correctly, because the reachability solvers were
kept from previous use.

Here we make sure the solvers are also cleared on a reset.