Add a prototype VMT frontend
VMT, Verification Modulo Theories, is a framework for working with
symbolic transition systems.
https://fm.fbk.eu/projects/verification-modulo-theories/
We add a prototype frontend that can parse a transition system defined
in VMT format, with the restriction that we can now deal only with
exactly one safety property.
Also, the language is restricted to linear arithmetic.