Termination: Fix symbolic condition on the step counter
On one hand, we can relax the condition, we don't have to require that
the counter starts at the value equal to maximum of program's variables.
On the other hand, we can make it more strict, by requiring that it is
greater than the maximum, not greater or equal.
Furthermore, we need to make sure we have some condition on the counter
even if there are no integer variables. Thus, we add the condition that
counter > 0.