TRL: Compute validity witness with quantifier elimination
When TRL proves system safe, it holds that the system extended with
learnt relations has a finite diameter d. That means that all states
reachable from Init are reachable within at most d steps of the
(extended) transition relation.
Thus we can write down the precise characterization of the reachable set
of states:
Reach(x) := \exists x_0,...,x_k
            Init(x) or
            Init(x_0) and Tr'(x_0, x) or
            Init(x_0) and Tr'(x_0, x_1) and Tr'(x_1,x) or
            .
            .
            .
            Init(x_0) and ... and Tr'(x_(d-1),x)

The set of all reachable states, Reach, is by definition an inductive
invariant of the system.
Using quantifier elimination we can obtain a quantifier free inductive
invariant. However, we need to eliminate O(dn) variables where d is the
diameter and n is the number of state variables of the system.